delete t/bugs/void_element.t - invalid test
This test file attempts to verify behaviour where tags are not closed; while
Zoom is designed primarily for HTML it specialises in XHTMLish HTML and as such
I don't really see a need to make this work.

If we were to support mangled HTML, we'd probably want to sit down and think
seriously about what rules to provide - we could inject missing close tag
events but I suspect that would require some bookkeeping to do sensibly.